+++ src/command_modules/azure-cli-cosmosdb/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.2.2
++++++
+* `cosmosdb create`: Add `--enable-multiple-write-locations` support.
+
 0.2.1
 +++++
 * Minor fixes